+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 dev:
+  - add "chain verify signedcontributionandproof"
   - show both block and body root in "block info"
 
 1.13.0: